Move service classes
- Create package for service classes and move them there
- Adjust scopes for (service) functions to not to break current functionality
