+++ src/test/java/org/phoebus/channelfinder/MetricsServiceIT.java
@@ -16,6 +16,7 @@ import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Property;
 import org.phoebus.channelfinder.entity.Tag;
 import org.phoebus.channelfinder.example.PopulateService;
+import org.phoebus.channelfinder.service.MetricsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -55,7 +56,8 @@ class MetricsServiceIT {
 
   @Autowired PopulateService populateService;
 
-  @Autowired MetricsService metricsService;
+  @Autowired
+  MetricsService metricsService;
 
   @Autowired private MockMvc mockMvc;
