+++ src/test/java/org/phoebus/channelfinder/epics/EpicsRPCRequestIT.java
@@ -21,6 +21,7 @@ import org.phoebus.channelfinder.TagManager;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Property;
 import org.phoebus.channelfinder.entity.Tag;
+import org.phoebus.channelfinder.service.ChannelFinderEpicsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.security.test.context.support.WithMockUser;
