+++ src/test/java/org/phoebus/channelfinder/performance/PopulateServiceIT.java
@@ -2,7 +2,7 @@ package org.phoebus.channelfinder.performance;
 
 import java.net.URL;
 import org.junit.jupiter.api.Test;
-import org.phoebus.channelfinder.AuthorizationService;
+import org.phoebus.channelfinder.service.AuthorizationService;
 import org.phoebus.channelfinder.example.PopulateService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
