+++ src/test/java/org/phoebus/channelfinder/processors/ChannelProcessorServiceTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.phoebus.channelfinder.entity.Channel;
+import org.phoebus.channelfinder.service.ChannelProcessorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Bean;
